humanoid docs wip
