+++ Documentation/lessons/non_physics_humanoids/images/humanoids/man_suit.jpg
