+++ Documentation/lessons/non_physics_humanoids/images/humanoids/smpl.gif
