+++ Documentation/lessons/non_physics_humanoids/images/humanoids/walk.gif
